chore: Changed how the get distance functs work
get_x_dist() functions now return an int boolean that determines
if an intersection occurs.
Standardised the code a little bit.
