+++ includes/objects.h
@@ -94,9 +94,9 @@ typedef struct s_minirt
 
 typedef struct s_intrsct
 {
-	t_coord	*r_vect;
+	t_coord	r_vect;
 	t_list	*obj_lst;
-	t_cam	camera;
+	t_coord	cam_pt;
 }	t_intrsct;
 
 #endif
